+++ data/src/main/java/team/retum/data/remote/datasource/declaration/AuthDataSource.kt
@@ -1,0 +1,12 @@
+package team.retum.data.remote.datasource.declaration
+
+import team.retum.data.remote.request.user.SendVerificationCodeRequest
+
+interface AuthDataSource {
+    suspend fun sendVerificationCode(sendVerificationCodeRequest: SendVerificationCodeRequest)
+
+    suspend fun verifyEmail(
+        email: String,
+        authCode: String,
+    )
+}
