+++ data/src/main/java/team/retum/data/remote/datasource/implementation/UserDataSourceImpl.kt
@@ -3,10 +3,7 @@ package team.retum.data.remote.datasource.implementation
 import team.retum.data.remote.api.UserApi
 import team.retum.data.remote.datasource.declaration.UserDataSource
 import team.retum.data.remote.request.user.SignInRequest
-import team.retum.data.remote.request.user.SendVerificationCodeRequest
-import team.retum.data.remote.request.user.SignUpRequest
 import team.retum.data.remote.response.user.SignInResponse
-import team.retum.data.remote.response.user.SignUpResponse
 import team.retum.data.storage.UserDataStorage
 import team.retum.data.util.HttpHandler
 import javax.inject.Inject
@@ -21,26 +18,6 @@ class UserDataSourceImpl @Inject constructor(
         .httpRequest {
             userApi.postLogin(
                 signInRequest = signInRequest,
-            )
-        }.sendRequest()
-
-    override suspend fun sendVerificationCode(
-        sendVerificationCodeRequest: SendVerificationCodeRequest,
-    ) = HttpHandler<Unit>()
-        .httpRequest {
-            userApi.sendVerificationCode(
-                sendVerificationCodeRequest = sendVerificationCodeRequest,
-            )
-        }.sendRequest()
-
-    override suspend fun verifyEmail(
-        email: String,
-        authCode: String,
-    ) = HttpHandler<Unit>()
-        .httpRequest {
-            userApi.verifyEmail(
-                email = email,
-                authCode = authCode,
             )
         }.sendRequest()
