+++ data/src/main/java/team/retum/data/repository/UserRepositoryImpl.kt
@@ -2,13 +2,8 @@ package team.retum.data.repository
 
 import team.retum.data.remote.datasource.declaration.UserDataSource
 import team.retum.data.remote.request.user.SignInRequest
-import team.retum.data.remote.request.user.toRequest
 import team.retum.data.remote.response.user.SignInResponse
-import team.retum.domain.param.user.CheckStudentExistsParam
 import team.retum.domain.param.user.SignInParam
-import team.retum.domain.param.user.SendVerificationCodeParam
-import team.retum.domain.param.user.SignUpParam
-import team.retum.domain.param.user.VerifyEmailParam
 import team.retum.domain.repository.UserRepository
 import javax.inject.Inject
 
@@ -39,25 +34,8 @@ class UserRepositoryImpl @Inject constructor(
         }
     }
 
-    override suspend fun sendVerificationCode(
-        sendVerificationCodeParam: SendVerificationCodeParam,
-    ) {
-        userDataSource.sendVerificationCode(
-            sendVerificationCodeRequest = sendVerificationCodeParam.toRequest(),
-        )
-    }
-
     override suspend fun fetchAutoSignInOption(): Boolean {
         return userDataSource.fetchAutoSignInOption()
-    }
-
-    override suspend fun verifyEmail(
-        verifyEmailParam: VerifyEmailParam,
-    ) {
-        userDataSource.verifyEmail(
-            email = verifyEmailParam.email,
-            authCode = verifyEmailParam.authCode,
-        )
     }
 
     private fun SignInParam.toRequest(): SignInRequest =
