+++ di/src/main/java/team/retum/di/NetworkModule.kt
@@ -11,6 +11,7 @@ import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 import team.retum.data.interceptor.AuthorizationInterceptor
 import team.retum.data.remote.api.ApplicationApi
+import team.retum.data.remote.api.AuthApi
 import team.retum.data.remote.api.BookmarkApi
 import team.retum.data.remote.api.BugApi
 import team.retum.data.remote.api.CodeApi
@@ -117,4 +118,8 @@ object NetworkModule {
     fun provideBugApi(
         retrofit: Retrofit,
     ): BugApi = retrofit.create(BugApi::class.java)
+
+    @Provides
+    @Singleton
+    fun provideAuthApi(retrofit: Retrofit): AuthApi = retrofit.create(AuthApi::class.java)
 }
