+++ domain/src/main/java/team/retum/domain/repository/AuthRepository.kt
@@ -1,0 +1,10 @@
+package team.retum.domain.repository
+
+import team.retum.domain.param.user.SendVerificationCodeParam
+import team.retum.domain.param.user.VerifyEmailParam
+
+interface AuthRepository {
+    suspend fun sendVerificationCode(sendVerificationCodeParam: SendVerificationCodeParam)
+
+    suspend fun verifyEmail(verifyEmailParam: VerifyEmailParam)
+}
