+++ domain/src/main/java/team/retum/domain/repository/UserRepository.kt
@@ -1,22 +1,10 @@
 package team.retum.domain.repository
 
-import team.retum.domain.param.user.CheckStudentExistsParam
 import team.retum.domain.param.user.SignInParam
-import team.retum.domain.param.user.SendVerificationCodeParam
-import team.retum.domain.param.user.SignUpParam
-import team.retum.domain.param.user.VerifyEmailParam
 
 interface UserRepository {
     suspend fun postLogin(
         param: SignInParam,
-    )
-
-    suspend fun sendVerificationCode(
-        sendVerificationCodeParam: SendVerificationCodeParam,
-    )
-
-    suspend fun verifyEmail(
-        verifyEmailParam: VerifyEmailParam,
     )
 
     suspend fun fetchAutoSignInOption(): Boolean
