+++ domain/src/main/java/team/retum/domain/usecase/user/VerifyEmailUseCase.kt
@@ -1,16 +1,17 @@
 package team.retum.domain.usecase.user
 
 import team.retum.domain.param.user.VerifyEmailParam
+import team.retum.domain.repository.AuthRepository
 import team.retum.domain.repository.UserRepository
 import javax.inject.Inject
 
 class VerifyEmailUseCase @Inject constructor(
-    private val userRepository: UserRepository,
+    private val authRepository: AuthRepository,
 ) {
     suspend operator fun invoke(
         verifyEmailParam: VerifyEmailParam,
     ) = kotlin.runCatching {
-        userRepository.verifyEmail(
+        authRepository.verifyEmail(
             verifyEmailParam = verifyEmailParam,
         )
     }
